requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
django==4.2.7; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
urllib3==2.1.0; python_version >= '3.8'
